Add more tests for createSarcohpagus
